+++ Sources/MapboxCoreNavigation/RerouteControllerDelegate.swift
@@ -1,0 +1,8 @@
+import Turf
+import MapboxDirections
+import MapboxNavigationNative
+
+
+protocol ReroutingControllerDelegate: AnyObject {
+    // TODO: fill with Native RerouteController and Alternative routes integration
+}